+++ src/apps/clients/apps.py
@@ -1,0 +1,10 @@
+"""Clients app config."""
+
+from django.apps import AppConfig
+
+
+class ClientsConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "apps.clients"
+    label = "clients"
+    verbose_name = "Clientes"